pyproject.toml
[project]
name = "spotify_autosync"
version = "0.1.0"
description = "Automates organization of new music in Spotify"
authors = [
    {name = "Prithvi Seri",email = "prithvi.seri@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "boto3 (>=1.40.21,<2.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.setuptools]
packages = ["spotify_autosync"]

[tool.setuptools.package-dir]
"" = "src"
